ch3

+++ os/src/task/task.rs
@@ -34,7 +34,7 @@ pub struct TaskInfo {
     /// Task status in it's life cycle
     pub status: TaskStatus,
     /// The numbers of syscall called by task
-    pub syscall_times: [usize; MAX_SYSCALL_NUM],
+    pub syscall_times: [u32; MAX_SYSCALL_NUM],
     /// Total running time of task
     pub time: usize,
 }
